Time: 3 ms (74.64%) | Memory: 87.6 MB (35.71%) - LeetSync

+++ average-waiting-time/average-waiting-time.java
@@ -9,7 +9,7 @@ class Solution {
             arrivalTime = customer[0];
             prepTime = customer[1];
             startTime = Math.max(startTime, arrivalTime);
-            double customerWait = (startTime + prepTime) - arrivalTime;
+            int customerWait = (startTime + prepTime) - arrivalTime;
             totalWait += customerWait;
             startTime += prepTime;
         }
